requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
python-dotenv==1.0.1

langchain==0.2.14
langchain-community==0.2.12
langchain-text-splitters==0.2.2
langgraph==0.2.34
langchain-core

chromadb==0.5.5
pypdf==4.3.1
pandas==2.2.2
openpyxl==3.1.5

# Gemini (LangChain integration)
langchain-google-genai==1.0.10
google-generativeai==0.7.2

streamlit==1.37.1
requests==2.32.3